+++ diesel_cli/src/infer_schema_internals/mysql.rs
@@ -92,6 +92,7 @@ pub fn determine_column_type(
     let unsigned = determine_unsigned(&attr.type_name);
 
     Ok(ColumnType {
+        sql_name: tpe.trim().to_lowercase(),
         rust_name: tpe.trim().to_camel_case(),
         is_array: false,
         is_nullable: attr.nullable,